+++ .gitignore
@@ -17,6 +17,7 @@ project/plugins/project/
 dynamodb-local/
 dynamodb-local-common/
 dynamodb-local-schedule-lambda/
+dynamodb-local-live-activities/
 
 # Scala-IDE specific
 .scala_dependencies
